added asCalendar attr
The datepicker can be used as an actual calendar. The code is checking for the attr asCalendar and if that exists then it don't hide the datepicker when clicking anything, but keeps it open for use as an calandar, im using this in my code tweeking the look of the datepicker with css. then it is up to the user if they wan't to hide the input element or not!